backend/requirements.txt
# core
fastapi>=0.68.0

# Postgres DBadapter
psycopg2==2.9.1

# .env support
python-dotenv>=0.19.0

# validation
pydantic[email]>=1.8.2

# ORM
sqlalchemy>=1.4.23

# hashing
passlib[bcrypt]>=1.7.4

# ASGI server
uvicorn==0.15.0
starlette>=0.14.2

# unit tests
pytest>=6.2.4
pytest-cov>=2.12.1
requests>=2.26.0

# for JWT
python-jose>=3.3.0

# for form data
python-multipart>=0.0.5
